// src/interfaces/spotify/spotify-artist.interface.ts
export interface Playability {
  playable: boolean;
  reason: Reason;
}
// ...
export interface TopTracksItem {
  uid: string;
  track: Track;
}

export interface Track {
  id: string;
  uri: string;
  name: string;
  playcount: string;
  discNumber: number;
  duration: Duration;
  playability: Playability;
  contentRating: ContentRating;
  artists: TrackArtists;
  album: Album;
}

export interface Album {
  uri: string;
  coverArt: AlbumCoverArt;
}

export interface AlbumCoverArt {
  sources: PurpleSource[];
}

export interface PurpleSource {
  url: string;
}

export interface TrackArtists {
  items: PurpleItem[];
}

export interface PurpleItem {
  uri: string;
  profile: UserLocation;
}

export interface UserLocation {
  name: string;
}

export interface ContentRating {
  label: ContentRatingLabel;
}

export enum ContentRatingLabel {
  Explicit = 'EXPLICIT',
  None = 'NONE',
}

export interface Duration {
  totalMilliseconds: number;
}